Project on the Movie Recommendation.
A project created on the movie recommendation, where the data is downloaded from the movielens100k website. By using the various data visualization tools visualize the data and drawn the conclusion based on the visualization, created a function where, first we get the user ratings of the particular movie. Then after,we need to get the series of the correlation data which are nears to the given movie. Then create the data frame and drop the null values, after that join the dataframe with the other dataframe of no.of users giving the ratings. At last we only get the values of the movies those which no.of people who rated are more than certain number of users based on the plot.